backend/python_ml/requirements.txt
fastapi
uvicorn[standard]
pymongo
scikit-learn
joblib
pandas
numpy
python-dotenv
